Backend
Trennung Wetter und run
